Separate solvers in differents files
